Added GetTileList() function to TileManager so that tiles can now be accessed, Buildings spawned through right-click now snap to the tile they're placed on, still need to set up system for getting adjacent tiles.
Added many explicit conversions in order to avoid warnings on strict compilers.

Changed base Object to render using "stRenderable" so that multiple meshes each using different textures (maybe shaders sometime too, although no real need for those yet) for things like health bars and selection boxes.

+++ include/Vector.h
@@ -30,11 +30,11 @@ public:
 
 	Vector Unitise();
 
-	float DotProduct(Vector left, Vector right);
+	double DotProduct(Vector left, Vector right);
 	
 	bool WithinBox(Vector a_vOtherVector, Vector a_vSize);
 
-	float Calculate2DAngle(Vector a_vOtherVector);
+	double Calculate2DAngle(Vector a_vOtherVector);
 
     Vector operator+(Vector a_Param);
 	Vector operator-(Vector a_Param);
